+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project adheres to Semantic Versioning.
 
-## [Unreleased]
+## [1.12.0] - 2026-07-05
 
 ### Added
 - Editable layout partials: `layout`, `header`, `footer`, and a new `styles`